Basic commands and variables work now!
Still need to make final_var work, and then add support
for functions, but overall, making good progress. After I get
final_var working, I should be completely done with the AliasConfig
file, which is turning out to be the largest chunk of the code.

The framework is set up for evaluating functions, so it shouldn't be hard
to add more functions later.