Added multiline construct and ivariable
In preperation for adding support for loops and other more advanced
constructs, added a construct that will allow a variable to be
defined on the right side. Also, made it easier for scripts to be written
by including the multiline construct so that users can put their scripts
on multiple lines. Also, I am in the process of reorganizing how functions
work, so that eventually I can add the ability to dynamically load function
modules, so that other plugin writers can extend the functionality of the
scripting language by adding their own functions to the language with ease.

Next is to figure out arrays.