Massive reorganization.
The compiler has been reorganized from the ground up, to split the alias engine into two parts.
The first part is the preprocessor, which pulls aliases out of the config file, and seperates the token
stream so that the internal portions (Minescripts) are seperate from the alias proper. Secondly, the
MinescriptCompiler class compiles the individual scripts. Breaking this up will allow Minescripts to
be run from locations other than an alias script. This is in anticipation of allowing Minescripts to
be triggered by events, such as players entering and exiting the game. Secondly, the evaluation
part of the code has been split up and can run functions seperately from the main server thread, if
possible, which is specified by the new runAsync method defined by the Function interface. While
this does have a few performance hits when running loops that interact with the bukkit api, overall
execution time should be shorter, and infinite loops will not freeze the server. Also, the compiler
will automatically insert sconcat() around arguments that are not seperated by commas. This eliminates
the need for the g() function, though the function still exists, for backwards compatibility reasons.

Several other functions have also been added. At this point, the code is compilable, but not
necessarily stable.