Assigning values to an array using a l-value approach is now possible.
This eliminates the need for array_set, though backwards compatibility has been maintained.
assign(@array[0], 'value') will work as expected now. Further, arrays are automatically created
as needed when using this syntax, so even if @array had been unassigned previously, it would now
be an array with 1 element in it. Documentation and a few more test cases are to follow.