Added a hard coded command that will fire off aliases.
This is meant to be used when you have another plugin that
fires of a CH alias, but it doesn't actually work. This is
likely due to the fact that the plugin isn't sending a
player command preprocess event, which is how CH reads commands.
So, this adds a hardcoded /runalias command, which takes the actual
alias as an argument. So if normally you run /aliascmd to fire off your
CH script, you would tell the plugin to run /runalias /aliascmd.