Made 2 important changes
First, all listeners are now converted into dirty listeners, if play-dirty mode is on. Only some events were being converted before, and it wasn't very efficien
t either. Now the list of listeners is walked through once, and all the listeners are converted, and from then on out, the event registration process is hijacke
d, so I can just do a conversion upon addition of an event, and not have to re-walk through the list to do the conversion. This means that the profiler should h
ave more correct data, and that the events themselves can be "super cancelled". Which brings me to...
Second, cancel()ing an event from within a bind handler works. This method cancels the bukkit event itself (if cancellable), as well as prevents the event from
being further propogated through the CH event framework. Also, the super_cancel() function is available.  To facilitate the ability to override plugins that don't behave properly, the listener injection framework is being made available to all events. This not only has the same effect as cancelling the event, but it also stops the event from being propagated further through the rest of the plugins (minus monitor level plugins).
Finally, event priorities, and delayed actions need to be added, and a specification for CH event bubbling needs to be developed in addition to the cancelling framework.