Added abstraction layer to allow easier migration away from bukkit if needed.
This commit may make the project extremely unstable. Builds in the near future should be regarded carefully, and should *only* be run on a test server first.

The changes here move 95% of the bukkit specific code into it's own abstraction package. Now, all the CH code only references the interfaces, which are implemented (right now) by classes that themselves reference the bukkit classes. In the future, should a new implementation be added, the CH specific code should not have to be touched, and multiple server implementations can easily be supported in parallel.