Changed the data type management stuff around.
THIS IS A BREAKING CHANGE. At least, potentially. If you used the is_boolean, is_integer, or is_double functions, your scripts *might* break. Previously, these functions returned whether or not a value was castable to the specified type. Now, they tell you if the data is *already* of the specified type. There was really previously no need for is_boolean, because it always returned true. All values can be used as a boolean, so there's no need to check that first, unless you actually care that it is of the boolean type. You may care if a string is numeric however, so to replace the old functionality of is_integer and is_double, you can now use is_numeric.

In addition, there are now casting methods, boolean(), integer(), double() and string(). These create new constructs of the appropriate type, which would then pass their corresponding is_* checks.