The chat event is run on the async thread now, if the underlying code doesn't need syncing.
This all happens transparently, but basically it means that user code will be optimized if possible, and
if not, it won't be, yet they didn't need to change anything at all. Additionally, prefilters are now compared
on the async thread, and if none match, the server thread is never acquired.