EXPERIMENTAL BUILD: SerializedPersistance now uses a more efficient, and more robust approach to saving.
Technical details: A byte[] is used as the data model, and when it is updated, it simply marks the data as dirty, at which point, it triggers the file writer to write to a temp file, then do a move operation. The idea is that the move operation is basically atomic, therefore creating a journaled file writer. Additionally, if the data model changes multiple times in between a write operation, this will not cause the intermediate versions of the data to be written out. This is both more memory and CPU efficient.