Fixed the file writer. This should fully fix persistance as well.
Previously, it was writing to a temp file, then moving that file, however, that appears to be too unreliable, so instead, it will write directly to the file, using a MappedByteBuffer, which *should* be more reliable, since it uses OS native mechanisms to actually do the mapping. If the OS supports journaling, it should be much less likely to get corrupted, but either way, it is still secure against inter-process corruption (and inter-thread corruption) and now, since the actual IO occurs on another thread, it should speed up things considerably.