Implemented Persistance Network. Additionally, refactored the environment.
This is a large change, and at this moment is a breaking change. Your persistance database will be empty for the time being, until I can get a import routine written. Additionally, I refactored how the environment works. Now, the environment is a very generic object, and a function must pull the appropriate sub environment out from the main environment. Additionally, it should declare in the @api tag that it uses a particular environment. This will eventually allow for fail-fast behavior when running a script externally. In the meantime, it is not specifically used, but all the functions should be implementing it properly thus far.