Added triml, trimr, split, and string_position.
Additionally, reg_split will now transform into split in certain cases, namely, where the regex used if fully literal (that is, the same behavior can be achieved with either function), and it is a constant expression. In my tests, split runs approximately twice as fast as reg_split, so it is favored in cases where regex is not strictly needed, however, it cannot be a replacement in all cases.