Better enums framework, enum updates, logging cleanup and fixing, other minor changes
The enum abstraction is always difficult due to enums being un-subclassable, or otherwise unusable in a dynamic way. Because of this, there is a need for custom enums that mirror the abstracted enums, and this is unfortunately hard to maintain, since enums can come and go from the concrete implementation, and the abstraction layer not know about it. So, at runtime now, when the server first starts up, it triggers a scan of all the registered enums for this platform. This has the benifit of alerting users that there is the potential for a problem, while also alertingto the fact that the abstracted enums are out of date. In *general* enums are 1:1, however, this framework provides a mechanism to abstract out the conversion to subclasses if needed.

Additionally, the logging mechanism has been updated and several bugs with it have been fixed. It should be much more reliable now. Startup messages are less verbose, while still containing the same information, and information that was incorrect has been corrected. 

Some biomes are no longer supported in vanilla minecraft, and these have been removed from the biome selections. A few new effects have been added as well, and should be available for use with the make_effect function. The documentation has been updated to reflect these enum changes.

A ReflectionUtils class has been added, which simplifies using reflection.

Each implementation can have it's own branding, which will eventually be phased in everywhere the "CommandHelper" name is used, and eventually, the Implementation static class will be replaced with instance based classes that provide certain functionality required for external access.