Added player_moved event.
This uses a secondary thread to do the lion's share of calculating player positions, and does not use bukkit's player move event directly. A "threshold" prefilter can be set, which makes this event more efficient than bukkit's event with native code, even, because it checks the prefilters on the secondary thread. The actual event is run on the server thread, if applicable, but this cost will have already been dedicated to processing the given code. The decision code is essentially free though, since it is run on an external thread. cancel() does not work yet, however, that will be implemented shortly.