In general, the brace syntax syntax works now.
It is now up to individual functions to implement the Braceable interface, which makes it eligible to support braces. Eventually keyword support will be added, which will allow for brace syntax for single keywords, like scope{ }. The final kinks need working out, then support for the other functions that support braces will be added.