After much thought, removed pfacing.
Since the yaw and pitch information is returned by ploc(), and the best option while keeping pfacing is to break backwards compatability anyways, it is probably best to just completely remove pfacing, and instruct users to simply use ploc() if they need to get that information. While set_pfacing is a duplication of set_ploc() in several ways, it is more convenient to use when setting, however it will be just as convenient to use ploc() as it would be to use pfacing, so there is no point in duplicating that code.

Additionally, this is the last compile error, so all the functions are now updated!