Added order consistent PropertiesManager class, and made the PN use it for the config file.
PropertiesManager works like java.util.Properties, except it extends LinkedHashMap<String, String>. This guarantees order, so it can reliably be determined which entry came first. The PN is supposed to work on this principal for conflicting entries that are equally specific, but because it was based on the Properties file, that couldn't be guaranteed. Now it can be, so this removes the warning (and the TODO).