Refined the unexpected error handler, and wrapped jedis exceptions.
Now, the unexpected error message handler will output the line of code properly if there are variables or strings in them. If there is a variable, it is printed out with the name, but the value that is currently assigned is printed out above it. This should make reproducing the errors much easier, as the code can almost be copy pasted now. Also, the JedisConnectionExceptions are wrapped into CRE.