Added http_request().
Finished the PublicSuffix security checking in the web utils, and broke them all into their own package. The WebUtility class is now complete, though it may need a bit of tweaking and bug fixing, but the primary use cases are working. The http_request function is also added now, and also appears to work.