Server starts up without error now, however, extensions aren't being loaded.
The startup time is still very slow however, so I'm going to continue to investigate to see if this can be improved at all. Changing over to not using ZipReader (in favor of the ZipIterator class) has exponentially increased the startup speed (from "unusably slow" to "noticeably slow"). Extensions aren't loading, but that's due to no classloaders being installed yet. I'll have to engineer a way for the extension manager to add classloaders to the ClassDiscovery class, like it was doing before, after I speed up the general class loading. I have temporarily disabled CheckOverrides and minimizeJar (in the shade plugin) to make builds faster, as they take several minutes to build currently, which is making testing very hard. The AbstractEnumVerification thread is not wired in appropriately to use the correct class addressing format apparently, as it is throwing an error too, but that is less important than the current issues.