Caching is now used, if available. The changes to ClassDiscovery should be complete now.
The only thing remaining is to make caches for each of the jars scanned be persistent, so
that the startup time will go much faster. The startup was nearly halved for the CH jar, so
I anticipate similar gains for the other jars, hopefully reducing the startup time (on my slow
computer) to < 10 seconds, down from the 20-30 seconds I was seeing.