Removed whitespace requirements in labels and slices.
Previously, labels were fixed so they didn't require particular whitespace,
but the code for that was placed in a compiler function. Information about
the token stream had been lost at that point, some of the information which
turned out to be vital, was lost during compilation. Now the full parsing
is done in the compiler, where it belongs.