Fixed MySQL Data Sources, updated docs.
Updating an existing value with MySQL was not working properly, 
only the first value was stored, updates were not being done.

MySQL is also being changed to use indexed tables, which
should be much more efficient than non-indexed tables,
and allows for use of the REPLACE INTO command, instead
of having to manually do a DELETE/INSERT or SELECT/UPDATE|INSERT
combination. This doubles the efficiency noted before. However,
anyone who has started to use the MySQL tables will have to
DROP TABLE `tableName`; in order for new queries to succeed.
No upgrade routine is being provided, since this is such a new
feature, and it was previously broken anyways.

Documentation has been updated in several places as well, and
a few changes and improvements to the built in syntax highlighter
have been made. Examples in the API page are now syntax highlighted.
Several more articles are now automatically managed, and a new article
describing the Persistence Network in more technical terms has been added.