Improper use of break is now a compile error.
Formerly, improper use would have been caught at runtime. Now
the rules are well defined, and cause a compile error. This requires
the break count to be hard coded, so break(@a) is not acceptable,
and causes a compile error. Breaks may not bubble up past
closures or procedures or the root of the code.