Finished capture_runas in ideal cases.
Plugins must use the provided CommandSender object in order for this to work, however. If they look up the player from the server via name, the proxied object will no longer be returned. This may be able to be fixed if play-dirty mode is active; the proxied CommandSender could be injected into the server's user table. Running as op currently doesn't work either, but that will be fixed in the future as well.