Convert CVoid and CNull to static getters.
This has the immediate effect of lessening the memory footprint, but the main
purpose of this is to prepare current and future code to be ready for if CNull
changes to a Java null, thus requiring less refactoring in the future. The memory
footprint effect will be minimal, but still advantageous. Additionally, code should
be neater. This does however potentially break extensions, which will need to
be recompiled, with the changes applied. This should prevent any future
breakages at this level in the future.