Update descriptions for /reloadaliases flags
Descriptions for some /reloadaliases flags are inconsistent with actual usage.

For example, "-r: Specifies that the persistence config file should be reloaded." however, -r prevents the persistence config file from being reloaded unless --whitelist is specified.

This contrasts with other flags whose descriptions are correct, i.e. "-s: Specifies that the scripts should not be reloaded."

Updated descriptions to be correct and consistent with actual usage and each other.