Refactored sk_region_create to work as intended.
Pre-existing uses of sk_region_create() are not affected by this change.

First it will search for a world defined in arg[0]. If no world is
found, it will search for a world in the LocationArrays. (If it finds
conflicting worlds an exception is thrown.) If no world is found, it
will get the commandSender’s world.

Bug fixed: LocationArrays needed to include a world else an exception
was thrown even if arg[0] defined a world. This was because
LocationArrays were converted into MCLocations and if a world was not
defined an error was thrown.

The function now directs through the Abstraction layer instead of
directly to Bukkit.