Added very basic support for variable type safety.
This does not add full support yet, for instance, in proc/closure definitions. This ONLY supports the trivial case of assign(). int @a = 'string', for instance is now a runtime error, as is trying to assign a string (or non-int) later in the code. Currently, these are runtime errors, though eventually it will be fully statically typed.