Added a new .msa only compiler.
This will eventually replace the non-pure mscript portions of the existing main compiler. This new lexer contains a simpler token list, as well as the ability to fully skip the pure mscript portions, which grants the advantage that the main compiler can completely ignore the msa portions, and can be fully decoupled from that syntax. The lexer is all that is complete now, however, a parser doesn't strictly need to be written. Remaining validation work may need to be done for edge cases, but the compiler appears to fully work in ideal cases.

New syntax is being supported for multiline start and end: <?ms and ?>, as well as lexer support for /** */ docblocks, which remain past compilation. This can be used in future expansions to provide description and other information for the command.