Fixed an issue where space mattered
Sometimes, the lexer would pick up whether or not a value is a decimal vs a concatenation
operator incorrectly. The whitespace shouldn't matter, but a decimal was assumed if the next character
was a number. This commit pushes the decision on whether or not it is a decimal into the later stage,
after whitespace has been removed. Now, "concatenating" two literal numbers will result in a double,
regardless of the whitespace before or after the dot.