Added a better error message for f(): label use
Making a label the result of dynamic values is not allowed. Previously, setting a label
such as: `f():' was caught by the compiler, but the error message was "Mismatched parenthesis"
because while it was caught, that was fairly accidental. Now, the catch is purposeful, and
the correct error message is displayed.