Made import/export accept IVar keys without cast.
Due to the removed deprecated export(@var) and import(@var), IVariables were not resolved before calling the exec. This caused the need to use export(string(@var), @value) and import(string(@var)). With this commit, it will accept IVariables and return an exception if the IVariable does not contain a CString.