Added ConfigRuntimeException.BuildException.
This is a intermediate change that allows a proper Exception extension system to be put in
place in the future, without causing old code to break. Ultimately, ideally one would use
new CREIOException, but in the meantime, until that system is in place, this will suffice.

Extensions should switch to this method as soon as possible.