Internationalize procedure names.
International letters, such as øæå and Здравствуйте work as procedure
names now. Other procedure restrictions still apply, such as requiring
the procedure to start with an underscore, and no other symbols in the
name, however international letters that are considered letters for the
purposes of Java's \p{L} character are now acceptable.