Added composite functions.
These will not be used in the core of MethodScript, however, may prove useful to end users
and extension authors. If it is easier to express a new function in terms of existing MethodScript
functions, then you can extend CompositeFunction to write a first class MethodScript function, that
runs other MethodScript. Besides the CompositeFunction class, a new test function,
test_composite_function() has been added to demonstrate usage. In general, it is more efficient
to write the code in native java, however if you would like to upgrade your procedures to a first
class function, you may do so with this mechanism. Extension authors may find this useful as well,
for creating utility methods based on existing functions.