Fixed PlayerOfflineEx in server_ping.
When modifying the player list in server_ping, a PlayerOfflineException
would be thrown if a player leaves between the time the player list for
the event was created and when the event was modified by modify_event()
in CommandHelper. Instead of checking which player should be on the
list, it now checks which players should NOT be on the list, fixing this
issue. Another noticable change is that players on the list that were
not online when the event was created will simply be ignored, where
previously a PlayerOfflineException was thrown.