Catch CancellationEx's when shutting down.
When the MC server shuts down and CommandHelper is disabled, scheduling
a task will result in a CancellationException. This Exception is now
caught and ignored or logged based on whether CommandHelper was still
enabled or not (before this commit, it was a raw stacktrace on the
console from a Timer thread).