Change the return of reg_match to array<string>
Previously, it returned array<mixed> by calling Static.resolveConstruct
on the returned values. This is bad for two reasons, one, because
Static.resolveConstruct is expensive, and two, the subtype of the array
components should not be dynamic. Additionally, reg_match_all already
returns array<string> so this was wrong on many levels. This *could*
have some issues with backwards compatibility, however, I think the
impact will be minimal.

Also, code cleanup in Regex.java