Improve block_break when drops are modified
There is no official way to modify drops from a BlockBreakEvent, so instead CH sets the block to air and spawns items. Previously protection plugins would have no say on whether CH could break a block when this was done. Also block loggers would see the block is AIR and ignore the event or record it inaccurately. This creates a new event so that other plugins can log or cancel the action accurately. It also lets lower priority binds see the modified drops.