Added x_get_oauth_token
Currently, only supports google, though others might work.

Also, it only supports in-browser authentication, but I intend on adding
a way to directly provide a refresh token, so that if you already have
that, you can truly automate the whole process. Also, it will prompt for
a new token every time, I still need to implement caching. I also intend
on moving the parameters into a profile so that secrets don't need to be
hardcoded in the code.