Fixed ArrayOutOfBoundsEx in the compiler.
When compiling "function(['bla'])" or similar code, the compiler would
attempt to convert the ['bla'] to "array_get(...)". Because the actual
array to get it from does not exist, it looks for this at index "-1",
causing an uncaught ArrayOutOfBoundsException. This commit adds a check
for this case and gives an expected compile exception instead.