Command block support and bugfixes.
Since CraftBukkit 1.12.1, execution of a CommandBlock fires a
ServerCommandEvent, allowing CommandHelper to listen to 'direct' aliases
(without /runalias prefix) in command blocks. This commit causes command
blocks without "/runalias" to be handled the same as the ones with
"/runalias" (effectively adding a "#" in front of the command block name
so player() will never return a valid player name for command blocks).
This commit also adds support for minecart command blocks in all places
where command block blocks are already supported.