Don't install profiler prefs multiple times or when an appropriate profiler exists already
In a couple cases the profiler preferences file was being written to twice when it should have been written to zero or one time. When compiling a file with a dozen procs in it, this would write to the profiler preferences file at least two dozen times. On a test server this previously accounted for 60% of total compile time. After this fix, generating a standalone environment now only accounts for 21% in the same test.