Add optional single slot parameter in set_pinv() (and optimizations)
Previously for every slot we set it would cause three packets to be sent to the player on the latest versions of Spigot. This is because we were updating the inventory to fix client syncing issues in the past. In theory this would be maybe 123 packets when setting a player inventory. This commit should reduce that to maybe 43 packets and maybe even fix some syncing when the inventory has other viewers. Because now after a function that alters an inventory is complete, it sends an update packet to all viewers of that inventory. This should ensure client syncing on all versions without doing too much extra work like before.