Change Minecraft reflection method to improve startup times
Despite only being used for sudo() right now, ClassDiscovery for the server jar adds up to 10 seconds to the startup time. The Bukkit convertor init also synchronizes during the ClassDiscovery process which may make things worse. This commit proposes that this extra time is not a necessary sacrifice and implements an alternative reflection method where needed.