Harmonize how void works
There is some formatting change, but the meat of the commit can be seen
in Static and BasicLogic. In Static, we change resolveConstruct to
return CVoid.VOID instead of CVoid.TYPE. Then, in sequals, we check
.typeof() not .getClass(). This is an important distinction, because we
want to base the equality on whether or not mscript says it's the same
type, not if java does. Actually, because of this change, the change to
Static isn't strictly necessary, but it is beneficial, since now we
standardize what we are returning.