Override methods in CSecureString that don't make sense to inherit
Also, needed to change array_get to work properly with ArrayAccess,
instead of assuming it is a string, if it is ArrayAccess and not a
CArray. array_normalize only makes sense with an actual array, so that
change to ArrayAccess was a bit premature. That change might actually
break some people's code, but whatever they were doing didn't make sense
in the first place, so you're welcome?