Improve core error logging.
Logging these errors is already slow, but the slow Main.loadSelfVersion() makes it take almost twice as long. This can cause noticeable tick lag. If this occurs with enough frequency the server will become unresponsive. So here we use the cached version from Bukkit when possible. The messaging is also improved a little.