BugFix for ambigous command detection.
Many commands were not detected as ambigous properly. Examples:
- *:/cmd2 $arg AND *:/cmd2 $arg
- *:/cmd11 $arg $arg2 AND *:/cmd11 $arg $arg2 [$arg3]
- *:/cmd14 arg AND *:/cmd14 $arg
- *:/cmd15 arg AND *:/cmd15 [$arg]

This commit will cause ambigous commands to throw a compile exception as
expected.
The previous code caught an IndexOutOfBoundsException, which's
fillInStacktrace() costed ~35% runtime spent in
AliasCore.LocalPackage.consoleMSA(). This has been replaced with a size
check to reduce runtime for MSA compiles by roughly this 35% (as tested
with my scripts and setup on my server with 254 defined aliases).