Create the shaded jar in addition to the non-shaded jar.
This allows CH to be independently included by other projects. The
biggest difference is that people now need to download the
commandhelper-version-SNAPSHOT-full.jar instead of the non -full
version. Cmdline users need to re-run install-cmdline to update the bash
command, as the name of the jar is encoded in the script.