Extra error management for event handlers.
When CH or a CH extension receives an event it cannot handle due to
being binary incompatible with a dependency, the error remained uncaught
and was passed to the server software that fired the event. This commit
makes sure those errors are caught and logged with useful additional
details for a bug report.