Add strict mode to preferences.ini. This allows a global setting.
Additionally, cleaned up the lexer code a bit, and made file options
higher priority than comments. Additionally, file options now have to
be the first non-comment, non-whitespace line in the file, but do not
have to be the VERY first thing in the file.

Whitespace changes, now that I have figured out how to use an IDE 
correctly, these changes should start to have to happen less and less.