Coercion of string "false" to boolean now provides a warning.
This is usually not what the user meant to do, as string "false"
evaluates to boolean true. Therefore, when we try to coerce a string
into a boolean, we first check to see if it is the string "false" and
if so, issue a warning. Eventually, I would like this to be configurable
per file, or at least at a more granular level, but anyways, for now 
this is a globally configurable warning in the logger-preferences.ini
settings.

This also required a change to the Static.getBoolean method, which has
been updated throughout the codebase to add a Target parameter. 
However, because extensions may
be using this method, it remains backwards compatible for now, but
the no Target method is deprecated and should be removed at next major
version change.