Add cmdline-install mechanism for Windows.
This installs a C# based launcher, which provides all the Windows
specific utilities to the system, but otherwise forwards all processing
on to the jar. This works much the same as the Linux launcher/bash
frontend. Unlike Linux, however, this is a much more complicated task,
and so there is a Visual Studios solution and everything included in
this commit.

There are still a few tasks remaining: Add the mscript.exe to the Path
so that it can be run from the cmd utility. Make the uninstall-cmdline
tool work. The prompt_* functions do not work yet.

Once these three things are done, this should be merged to master.