Added NonInheritImplements.
This allows classes to declare that they implement a specific interface,
but does not strictly require that subclasses also implement this
interface as well (through inheritance). There is a custom compile check
as well, but in general, this requires using the Helper class as well to
correctly cast to the interface and check instanceof.

The use case imagined here is for serialization and/or cloning, where
you want some classes in the heirarchy to be serializable or cloneable,
but not necessarily subclasses of those classes.