Add namespace support for native types
For the time being, there is no mechanism to add user classes,
or even using() statements, but now at least the mechanism is
fully defined, and the fully qualified class types are displayed
and internally tracked. The syntax for now is :: as the separator,
but it will probably change to . after some discussion and a
deprecation period. ms.lang and com.commandhelper are default
namespaces, and do not need to be specified, and in fact, at this
time cannot be specified, as there is currently no supported way
in code to fully qualify a class type. com.commandhelper will
eventually be removed once the CH functionality is split out, but
will be extensible by the embedding program, so will be added
back in specifically for CH scripts.

Before . can be used as the path separator, there needs to be a
deprecation period, for non-strict mode users, they could potentially
be concatenating bare strings, and that would interfere with this
mechanism, so this "feature" must be removed.