Add array_intersect, and others
Also added a default hashCode method to Construct, (as well as a very
thin hash() method that wraps that), which is used by array_intersect.
ArgumentValidation.getEnum is added as well. A -e option has been added
to the api command in the cmdline tools, which displays the examples for
a function. CArray now implements Iterable, but throws an exception for
associative arrays, as that doesn't make as much sense to happen.